rename getSigningKey to reserveSigningKey
