+++ packages/api/src/client/lexicons.ts
@@ -2811,29 +2811,6 @@ export const schemaDict = {
       },
     },
   },
-  ComAtprotoServerGetSigningKey: {
-    lexicon: 1,
-    id: 'com.atproto.server.getSigningKey',
-    defs: {
-      main: {
-        type: 'procedure',
-        description: 'Get a repo signing key for account creation.',
-        output: {
-          encoding: 'application/json',
-          schema: {
-            type: 'object',
-            required: ['signingKey'],
-            properties: {
-              signingKey: {
-                type: 'string',
-                description: 'Public signing key in the form of a did:key.',
-              },
-            },
-          },
-        },
-      },
-    },
-  },
   ComAtprotoServerListAppPasswords: {
     lexicon: 1,
     id: 'com.atproto.server.listAppPasswords',
@@ -2974,6 +2951,29 @@ export const schemaDict = {
             properties: {
               email: {
                 type: 'string',
+              },
+            },
+          },
+        },
+      },
+    },
+  },
+  ComAtprotoServerReserveSigningKey: {
+    lexicon: 1,
+    id: 'com.atproto.server.reserveSigningKey',
+    defs: {
+      main: {
+        type: 'procedure',
+        description: 'Reserve a repo signing key for account creation.',
+        output: {
+          encoding: 'application/json',
+          schema: {
+            type: 'object',
+            required: ['signingKey'],
+            properties: {
+              signingKey: {
+                type: 'string',
+                description: 'Public signing key in the form of a did:key.',
               },
             },
           },
@@ -7399,7 +7399,6 @@ export const ids = {
   ComAtprotoServerGetAccountInviteCodes:
     'com.atproto.server.getAccountInviteCodes',
   ComAtprotoServerGetSession: 'com.atproto.server.getSession',
-  ComAtprotoServerGetSigningKey: 'com.atproto.server.getSigningKey',
   ComAtprotoServerListAppPasswords: 'com.atproto.server.listAppPasswords',
   ComAtprotoServerRefreshSession: 'com.atproto.server.refreshSession',
   ComAtprotoServerRequestAccountDelete:
@@ -7409,6 +7408,7 @@ export const ids = {
   ComAtprotoServerRequestEmailUpdate: 'com.atproto.server.requestEmailUpdate',
   ComAtprotoServerRequestPasswordReset:
     'com.atproto.server.requestPasswordReset',
+  ComAtprotoServerReserveSigningKey: 'com.atproto.server.reserveSigningKey',
   ComAtprotoServerResetPassword: 'com.atproto.server.resetPassword',
   ComAtprotoServerRevokeAppPassword: 'com.atproto.server.revokeAppPassword',
   ComAtprotoServerUpdateEmail: 'com.atproto.server.updateEmail',
